==> 1_Data_to_spark.ipynb <==
import warnings
warnings.filterwarnings("ignore")

### Spark

from pyspark import init_spark
import pyspark.sql.functions as sf
from pyspark.sql.window import Window as sw
from pyspark.sql.types import StringType, IntegerType, ArrayType
spark = init_spark({"appName": 'to_spark'})

### Imports

import numpy as np
import pandas as pd
from matplotlib import pyplot as plt
from sklearn.metrics import roc_auc_score, precision_score

from tqdm import tqdm
import seaborn as sns
plt.style.use("bmh")
plt.rcParams['font.family'] = 'DejaVu Sans'

# Datasets

###### Targets train

target = pd.read_csv('./alfabattle2_train_target.csv')
spark.createDataFrame(target).write.format('orc').mode('overwrite').saveAsTable('alfa.andrey_auto_target_train')

###### Target test

target_test_content = pd.read_csv('./alfabattle2_test_target_contest.csv')
spark.createDataFrame(target_test_content).write.format('orc').mode('overwrite').saveAsTable('alfa.andrey_auto_target_test')

###### Train

train = pd.read_parquet(path='./train_transactions_contest/', engine='pyarrow')

%%time
possible_mccs = list(train.mcc.value_counts().keys())
current_mcc = possible_mccs[0]
c_train = train[train.mcc == current_mcc].reset_index(drop=True)
spark.createDataFrame(c_train).write.format('orc').mode('overwrite').saveAsTable('alfa.andrey_auto_train')
for current_mcc in tqdm(possible_mccs[1:]):
    c_train = train[train.mcc == current_mcc].reset_index(drop=True)
    spark.createDataFrame(c_train).write.format('orc').mode('append').saveAsTable('alfa.andrey_auto_train')

###### Test

test = pd.read_parquet(path='./test_transactions_contest/', engine='pyarrow')

possible_mccs = list(test.mcc.value_counts().keys())
current_mcc = possible_mccs[0]
c_test = test[test.mcc == current_mcc].reset_index(drop=True)
spark.createDataFrame(c_test).write.format('orc').mode('overwrite').saveAsTable('alfa.andrey_auto_test')
for current_mcc in tqdm(possible_mccs[1:]):
    c_test = test[test.mcc == current_mcc].reset_index(drop=True)
    spark.createDataFrame(c_test).write.format('orc').mode('append').saveAsTable('alfa.andrey_auto_test')

______

______

______

______

______

______

______

______

______

______

______
